Refactor MSDN Doc Provider into it's own task assembly and refactor task assembly compliation.
