+++ Directory.Build.targets
@@ -1,0 +1,3 @@
+<Project>
+  <Import Project="$(MSBuildThisFileDirectory)/SdkPackage.targets" Condition="'$(SharpGenSdkPackage)' == 'true'" />
+</Project>
