+++ Directory.build.props
@@ -5,7 +5,7 @@
     <VersionSuffix Condition="'$(CI)' == ''">local</VersionSuffix>
     <VersionSuffix Condition="'$(CI)' != '' AND '$(APPVEYOR_REPO_TAG_NAME)' == ''">ci$(APPVEYOR_BUILD_NUMBER)</VersionSuffix>
     <Authors>jkoritzinsky</Authors>
-    <Copyright>(c) 2010-2017 Alexandre Mutel, 2017 Jeremy Koritzinsky</Copyright>
+    <Copyright>(c) 2010-2017 Alexandre Mutel, 2017-2018 Jeremy Koritzinsky</Copyright>
     <PackageLicenseUrl>https://github.com/jkoritzinsky/SharpGenTools/blob/master/LICENSE.txt</PackageLicenseUrl>
     <PackageProjectUrl>https://github.com/jkoritzinsky/SharpGenTools</PackageProjectUrl>
     <Tags>SharpGen;CodeGen;CPlusPlus;COM</Tags>
