+++ SdkTests/SharpGen.Runtime/SharpGen.Runtime.csproj
@@ -3,12 +3,15 @@
   <PropertyGroup>
     <TargetFrameworks>netstandard1.1;netstandard2.0;net45</TargetFrameworks>
     <Description>Support classes for code generated by SharpGen.</Description>
-    <SignAssembly>true</SignAssembly>
+    <SignAssembly>false</SignAssembly>
     <SharpGenGenerateDoc>true</SharpGenGenerateDoc>
   </PropertyGroup>
 
   <ItemGroup>
     <SharpGenMapping Include="Mapping.xml" />
-    <PackageReference Include="System.Memory" Version="4.4.0-preview2-25405-01" Condition="'$(TargetFramework)' != 'net40'" />
+    <PackageReference Include="System.Memory" Version="4.4.0-preview2-25405-01" />
+    <PackageReference Include="SharpGen.Doc.Msdn.Tasks" Version="$(LocalPackageVersion)">
+        <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
   </ItemGroup>
 </Project>
