+++ SharpGen.Doc.Msdn.Tasks/SharpGen.Doc.Msdn.Tasks.props
@@ -1,0 +1,5 @@
+﻿<Project>
+  <PropertyGroup>
+    <DocProviderTargets>$(DocProviderTargets);DocumentCppModuleFromMsdn</DocProviderTargets>
+  </PropertyGroup>
+</Project>
