+++ SharpGen.Doc.Msdn/Connected Services/MTPS/Reference.cs
