+++ SharpGen.Doc.Msdn/MTPS.cs
