+++ SharpGen.Interactive/SharpGen.Interactive.csproj
@@ -11,7 +11,7 @@
     <OutputType>WinExe</OutputType>
     <RootNamespace>SharpGen.Interactive</RootNamespace>
     <AssemblyName>SharpGen.Interactive</AssemblyName>
-    <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.7</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
     <TargetFrameworkProfile />
@@ -80,6 +80,9 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Mono.Options" Version="5.3.0.1" />
+    <PackageReference Include="System.ValueTuple">
+      <Version>4.4.0</Version>
+    </PackageReference>
     <PackageReference Include="System.Xml.XmlDocument">
       <Version>4.3.0</Version>
     </PackageReference>
@@ -90,7 +93,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\SharpGen\SharpGen.csproj" />
+    <ProjectReference Include="..\SharpGen.Doc.Msdn\SharpGen.Doc.Msdn.csproj">
+      <Project>{4851ab71-078b-47c5-9e76-b7bff9162bcf}</Project>
+      <Name>SharpGen.Doc.Msdn</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\SharpGen\SharpGen.csproj">
+      <Project>{31f85a16-cb01-4456-be3c-76e9ff3a1343}</Project>
+      <Name>SharpGen</Name>
+    </ProjectReference>
   </ItemGroup>
   <Target Name="Pack" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
