+++ SharpGen/SharpGen.csproj
@@ -5,21 +5,15 @@
     <Description>Library for generating efficient C# code from C++ and COM headers.</Description>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="HtmlAgilityPack" Version="1.6.15" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="2.1.0" />
     <PackageReference Include="Microsoft.Win32.Registry" Version="4.3.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="9.0.0" />
     <PackageReference Include="System.Diagnostics.Process" Version="4.3.0" />
-    <PackageReference Include="System.Net.Requests" Version="4.3.0" />
     <PackageReference Include="System.Runtime.Extensions" Version="4.3.0" />
-    <PackageReference Include="System.ServiceModel.Duplex" Version="4.3.0" />
-    <PackageReference Include="System.ServiceModel.Http" Version="4.3.0" />
-    <PackageReference Include="System.ServiceModel.NetTcp" Version="4.3.0" />
-    <PackageReference Include="System.ServiceModel.Security" Version="4.3.0" />
-    <PackageReference Include="System.ValueTuple" Version="4.3.1" />
+    <PackageReference Include="System.ValueTuple" Version="4.3.0" />
     <PackageReference Include="System.Xml.XmlSerializer" Version="4.3.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <WCFMetadata Include="Connected Services" />
+    <PackageReference Include="System.Xml.XmlDocument" Version="4.3.0" />
+    <PackageReference Include="System.Xml.XPath.XmlDocument" Version="4.3.0" />
+    <PackageReference Include="System.Runtime.Serialization.Primitives" Version="4.3.0" />
+    <PackageReference Include="System.Runtime.Serialization.Xml" Version="4.3.0" />
   </ItemGroup>
 </Project>
