--- SharpGenTools.Sdk/_._
@@ -1,1 +1,0 @@
-﻿
