+++ SharpGenTools.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.27130.0
+VisualStudioVersion = 15.0.27130.2027
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharpGen", "SharpGen\SharpGen.csproj", "{31F85A16-CB01-4456-BE3C-76E9FF3A1343}"
 EndProject
@@ -13,6 +13,10 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharpGenTools.Sdk", "SharpGenTools.Sdk\SharpGenTools.Sdk.csproj", "{AB78CA3C-7463-456D-94FA-11B1942B757B}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharpGen.UnitTests", "SharpGen.UnitTests\SharpGen.UnitTests.csproj", "{13B246C9-F127-4AC5-8847-E9214C0ABD70}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharpGen.Doc.Msdn.Tasks", "SharpGen.Doc.Msdn.Tasks\SharpGen.Doc.Msdn.Tasks.csproj", "{108399D4-B687-4E7F-A60C-28EDC053A361}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharpGen.Doc.Msdn", "SharpGen.Doc.Msdn\SharpGen.Doc.Msdn.csproj", "{4851AB71-078B-47C5-9E76-B7BFF9162BCF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -96,6 +100,30 @@ Global
 		{13B246C9-F127-4AC5-8847-E9214C0ABD70}.Release|x64.Build.0 = Release|Any CPU
 		{13B246C9-F127-4AC5-8847-E9214C0ABD70}.Release|x86.ActiveCfg = Release|Any CPU
 		{13B246C9-F127-4AC5-8847-E9214C0ABD70}.Release|x86.Build.0 = Release|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Debug|x64.Build.0 = Debug|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Debug|x86.Build.0 = Debug|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Release|Any CPU.Build.0 = Release|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Release|x64.ActiveCfg = Release|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Release|x64.Build.0 = Release|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Release|x86.ActiveCfg = Release|Any CPU
+		{108399D4-B687-4E7F-A60C-28EDC053A361}.Release|x86.Build.0 = Release|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Debug|x64.Build.0 = Debug|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Debug|x86.Build.0 = Debug|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Release|x64.ActiveCfg = Release|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Release|x64.Build.0 = Release|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Release|x86.ActiveCfg = Release|Any CPU
+		{4851AB71-078B-47C5-9E76-B7BFF9162BCF}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
